requirements.txt
git+https://github.com/Algoritmi99/light.git
numpy~=2.2.3
overrides~=7.7.0
pandas~=2.2.3
opencv-python~=4.11.0.86
mediapipe~=0.10.21
PyYAML~=6.0.2